requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
Nuitka==1.8.4
ordered-set==4.1.0
protobuf==3.20.0
pycryptodome==3.19.0
pycryptodomex==3.19.0
pyperclip==1.8.2
PyYAML==6.0.1
requests==2.31.0
tabulate==0.9.0
urllib3==2.0.7
wasmer==1.1.0
wasmer_compiler_cranelift==1.1.0
xmltodict==0.13.0
zstandard==0.21.0
